Move BlurAccum trait into sealed module as supertrait of Primitive
- Move BlurAccum, U8Weight, and rounding_saturating_mul into the
  sealed module alongside PrimitiveSealed
- Add BlurAccum as a supertrait of Primitive, removing the need for
  explicit where bounds and #[expect(private_bounds)] on fast_blur
- Implement BlurAccum for all Primitive types (u8 with u32 accumulators,
  all others with f32)
- Rename acc_zero/acc_scale/acc_to_store to ZERO/scale/to_store
- Simplify rounding_saturating_mul bounds to just T: Primitive
